third_party/WebKit/LayoutTests/external/wpt/xhr/open-url-encoding-expected.txt

This is a testharness.js-based test.
PASS percent encode characters
FAIL lone surrogate assert_equals: expected "&%2365533;" but got "%26%2365533%3B"
Harness: the test ran to completion.
